Updates
Modified Code to include Jar outputs, also corrected a crash that occurs on Main after the pop up is displayed.

The game now runs smoother and I corrected a systemout print that needs to be manually enabled to not loop fast in the command line.

The Game also has a couple of beginner classes for the multiplayer functionality I'm adding soon.